+++ cmake/packaging.cmake
@@ -119,8 +119,8 @@ if( APPLE )
     configure_file(
      	${MRV2_DIR}/etc/macOS/mrv2.plist.in
      	${MRV2_BUNDLE_DIR}/Contents/Info.plist )
+
     
-    set(CMAKE_INSTALL_PREFIX "${CMAKE_BINARY_DIR}/install" CACHE PATH "Install prefix" FORCE)
     install(DIRECTORY ${MRV2_BUNDLE_DIR}
 	DESTINATION .
 	USE_SOURCE_PERMISSIONS
@@ -140,11 +140,11 @@ if( APPLE )
     # )
     
     # Copy the shell scripts into the bundles and make them executable
-    # configure_file(${MRV2_DIR}/etc/macOS/hdr.sh ${MRV2_BUNDLE_DIR}/Contents/MacOS/hdr)
+    # configure_file(${MRV2_DIR}/etc/macOS/hdr.sh ${HDR_BUNDLE_DIR}/Contents/MacOS/hdr)
     # 
     # configure_file(
     #  	${MRV2_DIR}/etc/macOS/hdr.plist.in
-    #  	${MRV2_BUNDLE_DIR}/Contents/Info.plist )
+    #  	${HDR_BUNDLE_DIR}/Contents/Info.plist )
     #
     #install(DIRECTORY ${HDR_BUNDLE_DIR}
     # 	DESTINATION .
@@ -154,18 +154,15 @@ if( APPLE )
 
     # Configure CPack for DragNDrop
     set(CPACK_GENERATOR "DragNDrop")
-    set(CPACK_PACKAGE_NAME "mrv2")
-    set(CPACK_PACKAGE_VERSION "1.0.0")
     set(CPACK_DMG_VOLUME_NAME "mrv2 Installer")
     set(CPACK_DMG_FORMAT "UDZO")
 
     # Set the volume icon
     set(CPACK_DMG_VOLUME_ICON ${MRV2_DIR}/etc/macOS/mrv2.icns)
     
-    set(CPACK_PACKAGE_FILE_NAME "mrv2-v${CPACK_PACKAGE_VERSION}")
-    set(CPACK_COMPONENTS_ALL "applications")
+    set(CPACK_COMPONENTS_ALL "macos_bundles")
     set(CPACK_INSTALL_CMAKE_PROJECTS "${CMAKE_BINARY_DIR};${CMAKE_PROJECT_NAME};applications;/")
-    set(CPACK_INSTALLED_DIRECTORIES "${CMAKE_INSTALL_PREFIX};.")
+    set(CPACK_INSTALLED_DIRECTORIES "${CMAKE_BINARY_DIR}/install;.")
 
 elseif(UNIX)